refactor(discord): extract utility functions from client into util.ts
Move the client-agnostic helpers (toFormData multipart serializer and
downloadAttachment CDN fetch) out of client.ts so the file is just the
photon-client API surface. Update inbound/media.ts to import
downloadAttachment from the new ./util module.
